Add files via upload
Welcome to the GoCaT2024 technical coding challenge. As a part of the AB Suite Tutorship team participated in this challenge, we're excited to present an educational video on the MOD feature implemented for AB Suite. This video serves as a valuable resource for enhancing our collective knowledge and proficiency with this feature.